Refactor test for joins on tables
- Merge repro for 12221 with the existing "base" test
- Merge repro for 13468 with the existing "base" test
